% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_card.R
\name{add_card}
\alias{add_card}
\title{Add cards to a deck}
\usage{
add_card(deck_path, front, back, difficulty = 1, source = "",
  source_url = "", rename_img = TRUE, max_pixels = 3e+05)
}
\arguments{
\item{deck_path}{The path to the deck folder}

\item{front}{The front of the cards to add}

\item{back}{The back of the cards to add}

\item{difficulty}{A positive number, typically between 1 and 10, indicating
the difficulty of the card. Lower difficulties will typically be presented
first.}

\item{source}{Where the information came from the copyright holder of images}

\item{source_url}{The location on the internet the infromation came from}

\item{rename_img}{If \code{TRUE}, rename the image files based on the
other side of the card.}

\item{max_pixels}{The maximum number of pixels an image can be. Larger images
will be shrunk to this size and saved as a jpeg.}

\item{return}{NULL}
}
\description{
Add cards to a deck
}
