% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_card.R
\name{add_image}
\alias{add_image}
\title{Add an image to a deck}
\usage{
add_image(img_path, name = basename(img_path), deck_path,
  max_pixels = 3e+05)
}
\arguments{
\item{img_path}{The path to the source image, either local or remote}

\item{name}{The name to save the image as. If the name already exists, a counter is added to the end. Spaces and capitals will be replaced.}

\item{deck_path}{The path to the deck directory}

\item{max_pixels}{The maximum number of pixels an image can be. Larger images
will be shrunk to this size and saved as a jpeg.}
}
\value{
The path to where the image was saved
}
\description{
Add an image to a deck
}
