% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{card_hash}
\alias{card_hash}
\title{Calculate a hash of a card}
\usage{
card_hash(content)
}
\arguments{
\item{content}{The content of either the front or the back of one or more cards.}
}
\value{
A hash string
}
\description{
Calculate a hash of the contents of a card.
}
\keyword{internal}
