% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_decks.R
\name{check_deck_format}
\alias{check_deck_format}
\title{Checks if deck data are formatted right}
\usage{
check_deck_format(decks, complain = TRUE)
}
\arguments{
\item{decks}{The paths to putative decks. Named by their file paths}

\item{complain}{If \code{TRUE}, issue warnings for any paths that do not
point to valid decks.}
}
\value{
A \code{TRUE/FALSE} vector corresponding to "paths".
}
\description{
Checks if deck data is formatted correctly
}
\keyword{internal}
