% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{combined_hash}
\alias{combined_hash}
\title{Get the combined hash for each card in a table}
\usage{
combined_hash(card_data)
}
\arguments{
\item{card_data}{A table with columns "front_hash" and "back_hash"}
}
\value{
A character vector of combined hashs
}
\description{
Get the combined hash for each card in a table
}
\keyword{internal}
