% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_deck_info}
\alias{get_deck_info}
\title{Read deck info}
\usage{
get_deck_info(path)
}
\arguments{
\item{path}{The file path to the deck folder}
}
\description{
Parse the information in a deck information YAML file
}
\keyword{internal}
