% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_deck_name}
\alias{get_deck_name}
\title{Get deck name}
\usage{
get_deck_name(deck_path)
}
\arguments{
\item{deck_path}{The path to the deck folder}
}
\description{
Get deck name from its path
}
\keyword{internal}
