% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{get_project_file}
\alias{get_project_file}
\title{Get the path to a project file}
\usage{
get_project_file(file_path, home = home)
}
\arguments{
\item{file_path}{File to find.}

\item{home}{The path to the folder containing the deck library, progress
file, and history file.}
}
\value{
vector
}
\description{
Get a file that might be a absolute path, in the current working directory, or in the home directory.
}
\keyword{internal}
