% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_deck}
\alias{init_deck}
\title{Create a deck template}
\usage{
init_deck(path, name = basename(path), description = "")
}
\arguments{
\item{path}{The path to new deck folder}

\item{name}{The name of the deck in a foramt that looks nice.}

\item{description}{The description of the deck. This would typically be a sentance or two.}
}
\value{
The path where the deck was created
}
\description{
Create a new empty deck
}
