% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_home_dir}
\alias{init_home_dir}
\title{Create a users home directory}
\usage{
init_home_dir(path)
}
\arguments{
\item{path}{The name/location of the deck library folder.}
}
\description{
Create a home directory with a deck library, history file, and progress file.
A user would generally have a single deck library and add decks to it. This
function creates a new deck library with an example deck to get started.
}
