% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{is_deck}
\alias{is_deck}
\title{Checks if decks folders are formatted right}
\usage{
is_deck(paths, complain = TRUE)
}
\arguments{
\item{paths}{The paths to putative decks}

\item{complain}{If \code{TRUE}, issue warnings for any paths that do not
point to valid decks.}
}
\value{
A \code{TRUE/FALSE} vector corresponding to "paths".
}
\description{
Checks if paths to putative decks point to decks formatted correctly.
}
\keyword{internal}
