% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_image_path}
\alias{is_image_path}
\title{Check if a string is a image path}
\usage{
is_image_path(path, test_url = curl::has_internet(), warn = TRUE)
}
\arguments{
\item{path}{The putative image path}

\item{test}{If \code{TRUE}, check that urls exists. This might take a second.}
}
\description{
Check if a string is a local or remote image path
}
\keyword{internal}
