% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_url}
\alias{is_url}
\title{Check for urls}
\usage{
is_url(text, test = FALSE)
}
\arguments{
\item{text}{The vector to test}

\item{test}{If \code{TRUE}, check that the url exists. This might take a second.}
}
\description{
Check if a vector contains urls
}
