% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{load_decks}
\alias{load_decks}
\title{Load one or more decks}
\usage{
load_decks(decks, library, user_dir, add_hash = TRUE)
}
\arguments{
\item{decks}{The names or folder paths of one or more decks to practice on.
By default, all decks in "library" are used.}

\item{library}{The path to the deck library. This is where the user stores
their decks and practice history. By default, this is a directory called
"deck_library" in the current working directory. If \code{NULL}, it will be
the current working directory.}

\item{add_hash}{If \code{TRUE}, add "front_hash" and "back_hash" columns
containing the md5 hash of the contents of the cards.}
}
\value{
A \code{data.frame}
}
\description{
Load one or more decks
}
\keyword{internal}
