% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{load_progress}
\alias{load_progress}
\title{Load practice progress}
\usage{
load_progress(progress, home = NULL, complain = TRUE,
  restrict_to_deck = NULL)
}
\arguments{
\item{progress}{The file used to store a user's practice progress for one or
more decks. By default, this is a file called "progress.csv" in the current
working directory.}

\item{complain}{If \code{TRUE}, issue warnings for any paths that do not
point to valid decks.}

\item{restrict_to_deck}{If a deck table is supplied, only return progress
data for that deck}
}
\value{
A \code{data.frame}
}
\description{
Load practice progress
}
\keyword{internal}
