% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{pick_card}
\alias{pick_card}
\title{Pick a card to practice}
\usage{
pick_card(deck, progress, focus = 0.5)
}
\arguments{
\item{deck}{A table containing deck information.}

\item{progress}{A table containing a users practice progress.}

\item{focus}{A number between 0 amd 1. Low numbers means less well known
cards will be practiced, whereas higher number means more well known cards
will be practiced.}
}
\value{
The row index in the deck of the card to practice
}
\description{
Pick a card to practice based on a user's practice progress.
}
\keyword{internal}
