% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounds.R
\name{play_sound}
\alias{play_sound}
\title{Play a sound}
\usage{
play_sound(name)
}
\arguments{
\item{The}{name of the sound file. Must be a wav file.}
}
\description{
Plays a sounds included with the pacakge in "inst/extdata/sounds"
}
\references{
Sound files used:

"wrong.wav": freesound.org user "Bertrof"
  https://freesound.org/people/Bertrof/sounds/131657/
"partial.wav: freesound.org user "JavierZumer"
  https://freesound.org/people/JavierZumer/sounds/257226/
"correct.wav: freesound.org user "LittleRainySeasons"
  https://freesound.org/people/LittleRainySeasons/sounds/335908/
}
\keyword{internal}
