% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{plot_card_side}
\alias{plot_card_side}
\title{Plot a card}
\usage{
plot_card_side(card, deck_path)
}
\arguments{
\item{card}{The content of the side of a card}
}
\description{
Plot a card. If it is a image file path, the contents of file is plotted.
}
\keyword{internal}
