% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{practice}
\alias{practice}
\title{Start a practice session}
\usage{
practice(home = getwd(), decks = NULL, progress = "progress.tsv",
  library = "decks", history = "history.tsv", tests = test_names(),
  record = TRUE, focus = 0.5, max_tests = 10)
}
\arguments{
\item{home}{The path to the folder containing the deck library, progress
file, and history file. If the folder was created using
\code{\link{init_home_dir}}, this is the only argument needed. By default,
this is the current working directory.}

\item{decks}{The names or folder paths of one or more decks to practice on.
If "library" is supplied, then these paths are relative to it. By default,
all decks in "library" are used.}

\item{progress}{The file used to store a user's progress for one or more
decks. By default, this is a file called "progress.tsv".  If "home" is
supplied, then this paths is relative to it. If \code{NULL}, no progress
file is used.}

\item{library}{The path to the deck library. This is where the user stores
their decks and practice history. By default, this is a directory called
"decks" in the "home" folder.  If "home" is supplied, then this paths is
relative to it.}

\item{history}{The file used to store a user's practice history for one or
more decks. By default, this is a file called "history.tsv".  If "home" is
supplied, then this paths is relative to it. If \code{NULL}, no history
file is used.}

\item{tests}{The names of test types to use in this practice session. By
default, the practice types specified by the decks are used.}

\item{record}{If \code{TRUE}, update the user's practice history with
the results of this practice session.}

\item{focus}{A number between 0 amd 1. Low numbers means less well known
cards will be practiced, whereas higher number means more well known cards
will be practiced.}

\item{max_tests}{The number of cards to present.}
}
\description{
Start practicing one or more decks.
}
