% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{present_test}
\alias{present_test}
\title{Present a test for a card}
\usage{
present_test(card, deck, tests = test_names(), progress = NULL)
}
\arguments{
\item{card}{The index of a card in "deck"}

\item{deck}{The data.frame containing the deck information}

\item{tests}{The name of tests to try to use}

\item{progress}{The progress table for the user}
}
\value{
The changes in scores resulting from the test
}
\description{
Pick a test ad present it to the user for a given card
}
\keyword{internal}
