% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{sample_learn_dist}
\alias{sample_learn_dist}
\title{Sample the learning distribution}
\usage{
sample_learn_dist(right, wrong, last_practiced = NULL)
}
\arguments{
\item{right}{The number or right answers/points.}

\item{wrong}{The number or wrong answers/points.}

\item{last_practiced}{TODO}
}
\value{
Integers between 1 and 0 for each input.
}
\description{
Uses the number of right and wrong answers and the time of the last practice
to sample from a beta distribution simulating the user's comprehension of the
card.
}
\keyword{internal}
