% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{save_history}
\alias{save_history}
\title{Save changes to the user's practice history}
\usage{
save_history(changes, history_path)
}
\arguments{
\item{changes}{A table containing the results of a practice session}

\item{history_path}{The path to the users history file}
}
\description{
Save changes to the user's practice history
}
\keyword{internal}
