% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{save_progress}
\alias{save_progress}
\title{Save the progress file}
\usage{
save_progress(progress, path)
}
\arguments{
\item{progress}{The current progress table}

\item{path}{Where to save the file}
}
\description{
Save the progress file
}
\keyword{internal}
