% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_answer}
\alias{test_answer}
\title{Typing test}
\usage{
test_answer(card, deck, progress, max_chars = 20, max_dist = 0.2,
  ignore_case = TRUE)
}
\arguments{
\item{card}{The index of the card to use}

\item{deck}{The table containing the deck information.}

\item{progress}{The progress table for the user}

\item{max_chars}{The maximum number of characters the answer side of a card
can have.}

\item{max_dist}{The maximum proportion of differences the answer can be to
the correct answer while still being considered correct.}

\item{ignore_case}{If \code{FALSE}, capitalization needs to be correct.}
}
\description{
Display one side of a card and ask which matches the other side of another
card shown. The user must type in the other side of the card.
}
\keyword{internal}
