% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_choose}
\alias{test_choose}
\title{Multiple choice test}
\usage{
test_choose(card, deck, max_choices = 4, pick_multiple = TRUE)
}
\arguments{
\item{card}{The index of the card to use}

\item{deck}{The table containing the deck information.}

\item{max_choices}{The maximum number of cards to show.}

\item{pick_multiple}{If \code{TRUE}, allow multiple correct answers}
}
\description{
Display one side of multiple cards and ask which matches the other side of
another card shown.
}
\keyword{internal}
