% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_review}
\alias{test_review}
\title{Review a card}
\usage{
test_review(card, deck)
}
\arguments{
\item{card}{The index of the card to use}

\item{deck}{The table containing the deck information.}
}
\description{
Review a card by flipping it over until moving to the next test.
}
\keyword{internal}
