% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{update_progress}
\alias{update_progress}
\title{Apply changes to the progress record}
\usage{
update_progress(changes, progress)
}
\arguments{
\item{changes}{A table containing the results of a practice session}

\item{progress}{The current progress table}
}
\value{
data.frame
}
\description{
Apply changes to the progress record
}
\keyword{internal}
