% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\docType{data}
\name{vocab}
\alias{vocab}
\title{Vocabulary}
\format{A data frame with two columns: Hebrew (the language tested) and English (the reference language)}
\usage{
vocab
}
\description{
Hebrew vocab words gathered over the course of the work day. Typically related to data science / fraud.
}
\keyword{datasets}
