% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_deck_name}
\alias{get_deck_name}
\title{Get a decks name}
\usage{
get_deck_name(path)
}
\arguments{
\item{path}{One or more paths to decks}
}
\value{
character
}
\description{
Get a decks name from its path
}
