% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_card.R
\name{is_image_path}
\alias{is_image_path}
\title{Check if a string is a image path}
\usage{
is_image_path(path, check_for_url = curl::has_internet(), warn = TRUE)
}
\arguments{
\item{path}{The putative image path}
}
\description{
Check if a string is a local or remote image path
}
\keyword{internal}
